f-footer@0.1.0 rename base vue component to be footer
